requirements.txt
beautifulsoup4==4.13.4
blinker==1.8.2
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
Flask==3.0.3
flask_cors==6.0.1
google-genai==1.46.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
